feat: folder always at top, move to user settings page
