+++ ui/src/RecentFilesDrawer/RecentFilesDrawer.tsx
@@ -14,19 +14,11 @@ import {
   Tooltip,
   Switch,
 } from "@chakra-ui/react";
-import {
-  useEffect,
-  useState,
-  useRef,
-  useCallback,
-  useContext,
-  ChangeEvent,
-} from "react";
+import { useEffect, useState, useRef, useCallback, useContext } from "react";
 import {
   workflowsTable,
   isFolder,
   foldersTable,
-  userSettingsTable,
 } from "../db-tables/WorkspaceDB";
 import {
   IconChevronDown,
@@ -77,12 +69,6 @@ export default function RecentFilesDrawer({ onClose, onClickNewFlow }: Props) {
     (window.localStorage.getItem(sortTypeLocalStorageKey) as ESortTypes) ??
       ESortTypes.RECENTLY_MODIFIED,
   );
-  const [folderOnTop, setFolderOnTop] = useState(false);
-  useEffect(() => {
-    userSettingsTable?.getSetting("foldersOnTop").then((res) => {
-      setFolderOnTop(res ?? false);
-    });
-  }, []);
 
   const loadLatestWorkflows = async () => {
     const all = (await workflowsTable?.listFolderContent()) ?? [];
@@ -125,18 +111,6 @@ export default function RecentFilesDrawer({ onClose, onClickNewFlow }: Props) {
     sortTypeRef.current = type;
     window.localStorage.setItem(sortTypeLocalStorageKey, type);
     !isFilter && setRefreshFolderStamp(Date.now());
-  };
-
-  const onFolderOnTopChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const state = e.target.checked;
-    userSettingsTable
-      ?.upsert({
-        foldersOnTop: state,
-      })
-      .then(() => {
-        setFolderOnTop(state);
-        !isFilter && setRefreshFolderStamp(Date.now());
-      });
   };
 
   useEffect(() => {
@@ -354,10 +328,7 @@ export default function RecentFilesDrawer({ onClose, onClickNewFlow }: Props) {
               </Menu>
               <Flex gap={2} align="center">
                 <Text>Folders on Top</Text>
-                <Switch
-                  isChecked={folderOnTop}
-                  onChange={onFolderOnTopChange}
-                />
+                <Switch />
               </Flex>
             </HStack>
             <SearchInput
