+++ ui/src/RecentFilesDrawer/WorkspaceSettingsModal.tsx
@@ -27,6 +27,7 @@ import { userSettingsTable } from "../db-tables/WorkspaceDB";
 import { ShortcutSettings } from "../settings/ShortcutSettings";
 import { validateOrSaveAllJsonFileMyWorkflows } from "../utils";
 import AutoSaveSettings from "../settings/AutoSaveSettings";
+import FolderOnTopSettings from "../settings/FolderOnTopSettings";
 
 export default function WorkspaceSettingsModal({
   onClose,
@@ -288,6 +289,7 @@ export default function WorkspaceSettingsModal({
                 <ShortcutSettings />
                 <AutoSaveSettings />
                 {/* <TwoWaySyncSettings /> */}
+                <FolderOnTopSettings />
               </VStack>
             </HStack>
           </ModalBody>
