+++ ui/src/db-tables/UserSettingsTable.ts
@@ -36,6 +36,7 @@ export class UserSettingsTable extends TableBase<UserSettings> {
       },
       autoSave: true,
       twoWaySync: false,
+      foldersOnTop: false,
     };
   }
